TimeTablePresenter is not longer necessary it is way more easy to do all the TimeTable stuff inside CoursePresenter as well. This is because Favourite Courses are the courses shown in TimeTable. TimeTable now shows Favourite Courses
